Moved Server and Console code together
