--- Main/Server/payloads/shell.sh
@@ -1,4 +1,0 @@
-#!/bin/sh 
-
-echo "Hello from the shell script"  >> ~/Desktop/ShellOutPutFile.txt
-echo "Hello there"
