+++ Main/ServerCode/Console/go.mod
@@ -1,0 +1,10 @@
+module console
+
+go 1.16
+
+require (
+	github.com/fatih/color v1.15.0 // indirect
+	github.com/gosuri/uitable v0.0.4
+	github.com/mattn/go-runewidth v0.0.14 // indirect
+	golang.org/x/term v0.7.0
+)
