+++ Main/ServerCode/Install_NiceC2.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+
+
+echo ███    ██ ██  ██████ ███████  ██████ ██████  
+echo ████   ██ ██ ██      ██      ██           ██ 
+echo ██ ██  ██ ██ ██      █████   ██       █████  
+echo ██  ██ ██ ██ ██      ██      ██      ██      
+echo ██   ████ ██  ██████ ███████  ██████ ███████ 
+                                             
+
+sh Server/install.sh
+
+sh Console/install.sh
