+++ Main/ServerCode/Server/go.mod
@@ -1,0 +1,3 @@
+module server
+
+go 1.16
