+++ Main/ServerCode/Server/nodes.json
@@ -1,0 +1,9 @@
+[
+  {
+    "ID": "FCB85CB9-9452-539B-9988-48A4C5E3DFD3",
+    "Hostname": "Jasper's Air",
+    "Platform": "darwin",
+    "First_Check_In": "2023-04-08 21:52:47.091856 +0100 BST m=+0.495105168",
+    "Last_Check_In": "2023-05-05 23:54:39.263786 +0100 BST m=+447.488345751"
+  }
+]
