+++ Main/ServerCode/Server/payloads/PowerShell.ps1
@@ -1,0 +1,2 @@
+"This was written by a PS script" | Out-File -FilePath ~\Desktop\PSout.txt
+echo "Hello"
