+++ Main/ServerCode/Server/payloads/shell.sh
@@ -1,0 +1,4 @@
+#!/bin/sh 
+
+echo "Hello from the shell script"  >> ~/Desktop/ShellOutPutFile.txt
+echo "Hello there"
